Replaced vectormath library with mgl32

+++ controller/controller.go
@@ -5,6 +5,6 @@ type Controller interface {
 	BindKeyAction(function func(), key Key, action Action)
 	SetMouseAction(function func(button MouseButton, action Action))
 	BindMouseAction(function func(), button MouseButton, action Action)
-	BindAxisAction(function func(xpos, ypos float64))
-	BindScrollAction(function func(xoffset, yoffset float64))
+	BindAxisAction(function func(xpos, ypos float32))
+	BindScrollAction(function func(xoffset, yoffset float32))
 }
